Refactor import code.
